requirements-dev.txt
# Development dependencies for EC2 Patching Workflow
pytest>=7.0.0
pytest-cov>=4.0.0
moto>=4.2.0
boto3-stubs>=1.28.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.0
pre-commit>=3.0.0
# Note: markdownlint-cli2 is an npm package; if needed, install via Node separately.
